Fix most EmbeddedEditor tests
